add test for config file

+++ scripts/make-sb.sh
@@ -42,35 +42,37 @@ function parse_yaml_for_vhosts {
    }'
 }
 
-# Parse the local config yml file into the global vars.
-eval $(parse_yaml_for_vhosts ${CONFIG_FILE})
-#( set -o posix ; set ) | more
-for vhost in ${!apache_vhosts__documentroot*}
-do
-    # Get the docroot directory name.
-    directory=${!vhost}
-    directory=${directory/__/\/}
-    if [ ! -d sites/$directory ]; then
-        echo "Type the branch name that you want to check out into the directory $directory, followed by [ENTER]:"
-        read branch
-        cd build
-        git checkout $branch
-        cd ../
-        # Check to see if the developer make file is available.
-        if [ -f build/springboard-developer.make ]; then
-            drush make --no-gitinfofile build/springboard-developer.make sites/$directory;
-        else
-            # Use the standard make file.
-            drush make --no-gitinfofile --working-copy build/springboard-mtsb.make sites/$directory;
+if [ -f CONFIG_FILE ]; then
+    # Parse the local config yml file into the global vars.
+    eval $(parse_yaml_for_vhosts ${CONFIG_FILE})
+    #( set -o posix ; set ) | more
+    for vhost in ${!apache_vhosts__documentroot*}
+    do
+        # Get the docroot directory name.
+        directory=${!vhost}
+        directory=${directory/__/\/}
+        if [ ! -d sites/$directory ]; then
+            echo "Type the branch name that you want to check out into the directory $directory, followed by [ENTER]:"
+            read branch
+            cd build
+            git checkout $branch
+            cd ../
+            # Check to see if the developer make file is available.
+            if [ -f build/springboard-developer.make ]; then
+                drush make --no-gitinfofile build/springboard-developer.make sites/$directory;
+            else
+                # Use the standard make file.
+                drush make --no-gitinfofile --working-copy build/springboard-mtsb.make sites/$directory;
+            fi;
+            cd sites/$directory;
+            echo sites/all >> .gitignore; echo profiles/sbsetup >> .gitignore;
+            cd ../../
+            # Create a sustainer.key file in sites/default/files
+            mkdir -p sites/$directory/sites/default/files
+            if [ ! -e sites/$directory/sites/default/files/sustainer.key ]; then
+              echo $directory > sites/$directory/sites/default/files/sustainer.key
+            fi;
         fi;
-        cd sites/$directory;
-        echo sites/all >> .gitignore; echo profiles/sbsetup >> .gitignore;
-        cd ../../
-        # Create a sustainer.key file in sites/default/files
-        mkdir -p sites/$directory/sites/default/files
-        if [ ! -e sites/$directory/sites/default/files/sustainer.key ]; then
-          echo $directory > sites/$directory/sites/default/files/sustainer.key
-        fi;
-    fi;
 
-done
+    done
+fi
